refactor components, smooth scroll chat
