+++ frontend/src/components/views/index.ts
@@ -2,8 +2,7 @@ import { PreGameHostView } from "./pre-game/pre-game-host-view";
 import { PreGamePlayerView } from "./pre-game/pre-game-player-view";
 import { PickingDrawerView } from "./playing/picking/picking-drawer-view";
 import { PickingGuesserView } from "./playing/picking/picking-guesser-view";
-import { DrawingDrawerView } from "./playing/drawing/drawing-drawer-view";
-import { DrawingGuesserView } from "./playing/drawing/drawing-guesser-view";
+import { DrawingView } from "./playing/drawing/drawing-view";
 import { PostDrawingView } from "./playing/post-drawing-view";
 import { EnterCodeView } from "./join-room/enter-code-view";
 import { EnterPlayerInfoView } from "./join-room/enter-player-info-view";
@@ -15,7 +14,6 @@ export {
 	PreGamePlayerView,
 	PickingDrawerView,
 	PickingGuesserView,
-	DrawingDrawerView,
-	DrawingGuesserView,
+	DrawingView,
 	PostDrawingView,
 };
