+++ frontend/src/components/views/playing/picking/picking-guesser-view.tsx
@@ -1,7 +1,7 @@
 import { generateAvatar } from "@/lib/avatar";
 import { useSelector } from "react-redux";
 import { RootState } from "@/state/store";
-import { getPickingPlayer } from "@/lib/player";
+import { getDrawingPlayer } from "@/lib/player";
 import { SkyScene } from "@/components/scenes/sky-scene";
 import { BobbingDoodle } from "@/components/doodle/bobbing-doodle";
 import { AnimatePresence } from "framer-motion";
@@ -9,18 +9,18 @@ import { AirplaneDoodle } from "@/components/doodle/airplane-doodle";
 
 export function PickingGuesserView() {
 	const players = useSelector((state: RootState) => state.room.players);
-	const pickingPlayer = getPickingPlayer(players);
+	const drawingPlayer = getDrawingPlayer(players);
 	const isFirstPhase = useSelector(
 		(state: RootState) => state.game.isFirstPhase
 	);
-	if (!pickingPlayer) return null;
-	const avatarSvg = generateAvatar(pickingPlayer.avatarSeed);
+	if (!drawingPlayer) return null;
+	const avatarSvg = generateAvatar(drawingPlayer.avatarSeed);
 
 	return (
 		<SkyScene>
 			<img src={avatarSvg} className="w-20 h-20 rounded-lg shadow-accent" />
 			<h1 className="text-2xl lg:text-3xl font-bold px-8 lg:px-0 text-center">
-				{pickingPlayer.name} is picking a word to sketch
+				{drawingPlayer.name} is picking a word to sketch
 			</h1>
 			<AnimatePresence>
 				<BobbingDoodle
