+++ frontend/src/config/spring.ts
@@ -1,0 +1,14 @@
+export const containerSpring = {
+	type: "spring",
+	stiffness: 350,
+	damping: 45,
+	mass: 1,
+	restDelta: 0.01,
+};
+export const containerSpringFast = {
+	type: "spring",
+	stiffness: 650,
+	damping: 45,
+	mass: 1,
+	restDelta: 0.01,
+};
