+++ frontend/src/main.tsx
@@ -9,7 +9,7 @@ import { TermsOfService } from "./pages/terms-of-service";
 import { PrivacyPolicy } from "./pages/privacy-policy";
 import { HowToPlay } from "./pages/how-to-play";
 import { PersistGate } from "redux-persist/integration/react";
-import { UIHeader } from "./components/header";
+import { UIHeader } from "./components/ui-header";
 import { SoundProvider } from './providers/sound-provider';
 
 createRoot(document.getElementById("root")!).render(
