+++ frontend/src/pages/how-to-play.tsx
@@ -1,4 +1,4 @@
-import { ContentLayout } from "@/components/content-layout";
+import { ContentLayout } from "@/layouts/content-layout";
 
 export function HowToPlay() {
 	return (
