+++ frontend/src/pages/terms-of-service.tsx
@@ -1,4 +1,4 @@
-import { ContentLayout } from "@/components/content-layout";
+import { ContentLayout } from "@/layouts/content-layout";
 
 export function TermsOfService() {
 	return (
